[frontend] Simplify dimension selectors
This patch moves the dimension selectors to their own component and
simplifies the HTML considerably.
